utils: add `TempWriter` and remove from `hash-ir-utils`
