+++ compiler/hash-utils/src/lib.rs
@@ -19,6 +19,7 @@ pub mod range_map;
 pub mod scoping;
 pub mod stack;
 pub mod state;
+pub mod temp_writer;
 pub mod timing;
 pub mod tree_writing;
